dashboard/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pandas==2.1.3
plotly==5.18.0
